welcome js updated for modularity, works
