+++ welcome.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
+    <title>Snow in Spring</title>
 </head>
 <body>
     <article class="character_creation">
@@ -35,6 +35,6 @@
             <div class="character_container"></div>
         </section>
     </article>
-    <script type="module" src="./scripts/stats.js"></script>
+    <script type="module" src="./scripts/welcome.js"></script>
 </body>
 </html>
